requirements.txt
Flask
gunicorn
opencv-python-headless
Pillow